Maintenance: Update welcome screen
Updating references to Corona to Solar2D and updating the copyright disclaimer's year.

Also removing downloadFeeds function as unnecessary. The website that the function checks does not exist anymore and there will more than likely never be any feed for it to download again, so it's just an unnecessary network call that will never succeed.